+++ src/main/java/diploma/geometry/Vector.java
@@ -2,7 +2,7 @@ package diploma.geometry;
 
 public class Vector extends Tuple {
 
-    public Vector(float x, float y, float z) {
-        super(x, y, z, 0.0f);
+    public Vector(double x, double y, double z) {
+        super(x, y, z, 0.0);
     }
 }
